requirements.txt
airflow-code-editor
black
fs-s3fs
fs-gcsfs
lxml
beautifulsoup4
